Incorporate spam checking into people profile editing.  Move the edit class where these overrides are done from plugins/upstream/addbook.py to plugins/upstream/code.py since all page edits seem to use that class.
