+++ openlibrary/plugins/upstream/code.py
@@ -11,6 +11,7 @@ from six import PY3
 import web
 
 from infogami import config
+from infogami.core import code as core
 from infogami.infobase import client
 from infogami.utils import delegate, app, types
 from infogami.utils.view import public, safeint, render
@@ -20,6 +21,7 @@ from infogami.utils.context import context
 from openlibrary import accounts
 
 from openlibrary.plugins.upstream import addbook, covers, merge_authors, models, utils
+from openlibrary.plugins.upstream import spamcheck
 from openlibrary.plugins.upstream import borrow, recentchanges  # TODO: unused imports?
 from openlibrary.plugins.upstream.utils import render_component
 
@@ -31,6 +33,28 @@ class static(delegate.page):
     def GET(self):
         host = 'https://%s' % web.ctx.host if 'openlibrary.org' in web.ctx.host else ''
         raise web.seeother(host + '/static' + web.ctx.path)
+
+
+class edit(core.edit):
+    """Overwrite ?m=edit behaviour for author, book, work, and people pages."""
+    def GET(self, key):
+        page = web.ctx.site.get(key)
+
+        if web.re_compile('/(authors|books|works)/OL.*').match(key):
+            if page is None:
+                raise web.seeother(key)
+            else:
+                raise web.seeother(page.url(suffix="/edit"))
+        else:
+            return core.edit.GET(self, key)
+
+    def POST(self, path):
+        if web.re_compile('/(people/[^/]+)').match(path) and spamcheck.is_spam():
+            return render_template('message.html', 'Oops',
+                                   'Something went wrong. Please try again later.')
+        return core.edit.POST(self, path)
+
+
 
 # handlers for change photo and change cover
